@@ -7,15 +7,13 @@ ext {
 }
 
 android {
-    compileSdkVersion 27
+    compileSdk 33
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 4
-        versionName "1.3"
+        targetSdkVersion 33
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
 
     buildTypes {
@@ -23,24 +21,31 @@ android {
             minifyEnabled false
         }
     }
-    lintOptions {
+    namespace 'com.shagi.materialdatepicker'
+    lint {
         abortOnError false
     }
-
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
+    }
 }
 
 dependencies {
-    implementation 'com.android.support:support-v4:27.0.2'
-    implementation 'com.android.support:support-v13:27.0.2'
-    implementation 'com.android.support:design:27.0.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
+    implementation 'com.google.android.material:material:1.9.0'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
 
-    testImplementation 'junit:junit:4.12'
-    testImplementation 'com.pholser:junit-quickcheck-core:0.7'
-    testImplementation 'com.pholser:junit-quickcheck-generators:0.7'
+    testImplementation 'junit:junit:4.13.2'
+    testImplementation 'com.pholser:junit-quickcheck-core:1.0'
+    testImplementation 'com.pholser:junit-quickcheck-generators:1.0'
 
-    androidTestImplementation 'com.android.support:support-annotations:27.0.2'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test:rules:1.0.1'
+    androidTestImplementation 'androidx.annotation:annotation:1.6.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test:rules:1.5.0'
+
+    implementation 'androidx.recyclerview:recyclerview:1.3.0'
 }
 
-apply from: 'https://raw.githubusercontent.com/blundell/release-android-library/master/android-release-aar.gradle'
+//apply from: 'https://raw.githubusercontent.com/blundell/release-android-library/master/android-release-aar.gradle'
