@@ -17,7 +17,7 @@
 package com.shagi.materialdatepicker.date;
 
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.Calendar;
 
