@@ -18,8 +18,8 @@ package com.shagi.materialdatepicker.date;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.shagi.materialdatepicker.Utils;
 
